feat: add verified jws representation to serializeTransaction (#36)
Possible fix to #34 

As stated in the StoreKit
[documentation](https://developer.apple.com/documentation/storekit/transaction#Verify-transactions):

> To perform your own validation on the device, use the verification
result’s jwsRepresentation string, and use the provided convenience
properties headerData, payloadData, and signatureData. For added control
and security, send the jwsRepresentation to your server to verify.

However, the current return value of the library's `requestPurchase`
only returns the `jsonRepresentation` of the `Transaction` object, and
not `jwsRepresentation` of the verification result, which makes it more
complicated to implement a server-side verification.

With this change, user can correctly get the required `JWS` that they
can use to implement their own verification during the `purchase` flow.